Add files via upload
Challenge Description
Participating in UMA governance is one of the most important aspects of being a contributing community member. It is one of the pillars of UMA and that is why we’d like to see the voters get rewarded with an awesome “I voted” POAP! As long as you follow the submission requirements you can be as creative as possible. Quick note our community likes chickens, eggs, legos, and superhero capes!

Submission Requirements
Must have the official UMA logo letters
Must have the I voted text or similar text such as ‘voted’